Switch run-template-extractor command to run from v2 instead of v1
