+++ packages/project-builder-server/src/template-extractor/index.ts
@@ -1,2 +1,1 @@
-export * from './run-template-extractor-v2.js';
 export * from './run-template-extractor.js';
